LPS-54105 moves java
