+++ modules/apps/trash/trash-web/src/com/liferay/trash/web/search/EntryDisplayTerms.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.portlet.trash.search;
+package com.liferay.trash.web.search;
 
 import com.liferay.portal.kernel.dao.search.DisplayTerms;
 import com.liferay.portal.kernel.util.ParamUtil;
